requirements.txt
streamlit==1.36.0
openai>=1.40.0
Pillow>=10.3.0
python-dotenv>=1.0.1